Fix build with panic="abort"
When building with `panic="abort"` and `std` built from source, the size
of `Mutex<()>` varies when wrapped in `Option<>` or not such that the
owned and loaned types have different sizes.

This is specifically due to the `poison` flag in `Mutex` getting
compiled out when the `panic` mode is not set to `"unwind"`.

The build fails as follows:

```
note: evaluation panicked: Size mismatch: type z_loaned_mutex_t has size 16 while type z_owned_mutex_t has size 24
```

To address, we switch to `decl_c_type_inequal!` which seems to be safe
since we shouldn't be `reinterpret_cast()`-ing this type and `zenoh-cpp`
doesn't seem to ever reference it.